requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
Pillow>=10.0.0
gunicorn==22.0.0
whitenoise==6.6.0
drf-yasg==1.21.7
dj-database-url==2.1.0
psycopg2-binary==2.9.9
